Moved `like` / `unlike` API actions to dedicated controller
ref https://linear.app/ghost/issue/AP-1082

Moved `like` / `unlike` API actions to a dedicated controller to follow the
same pattern as the `block` controller, in preparation for modification of the
`handleLike` method to prevent processing of requests from blocked accounts
